refactor: renamed to viewer
