+++ jumanji/environments/packing/knapsack/viewer.py
