+++ jumanji/environments/routing/cvrp/viewer.py
