+++ jumanji/environments/routing/maze/env.py
@@ -24,9 +24,9 @@ from jumanji import specs
 from jumanji.env import Environment
 from jumanji.env_viewer import Viewer
 from jumanji.environments.routing.maze.constants import MOVES
-from jumanji.environments.routing.maze.env_viewer import MazeEnvViewer
 from jumanji.environments.routing.maze.generator import Generator, RandomGenerator
 from jumanji.environments.routing.maze.types import Observation, Position, State
+from jumanji.environments.routing.maze.viewer import MazeEnvViewer
 from jumanji.types import TimeStep, restart, termination, transition
 
 
@@ -85,7 +85,7 @@ class Maze(Environment[State]):
         generator: Optional[Generator] = None,
         time_limit: Optional[int] = None,
         render_mode: str = "human",
-        viewer: Optional[Viewer] = None,
+        viewer: Optional[Viewer[State]] = None,
     ) -> None:
         """Instantiates a `Maze` environment.
