+++ jumanji/environments/routing/tsp/viewer.py
